backend/text_to_sql/requirements.txt
# python==3.13
# conda activate generate

google-generativeai
python-dotenv
pg8000
sqlalchemy